bidgala/requirements.txt
aftership==1.1.0
aioredis==1.3.1
arabic-reshaper==2.1.1
asgiref==3.3.0
astroid==2.4.2
async-timeout==2.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.1
boto3==1.14.33
botocore==1.17.33
cached-property==1.5.2
certifi==2020.4.5.1
#cffi==1.15.0
cffi==1.14.5
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
constantly==15.1.0
coverage==5.3
coveralls==2.1.2
cryptography
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.4.0
django-livereload-server==0.3.2
django-rest==0.8.7
django-searchable-encrypted-fields==0.1.9
django-session-timeout==0.1.0
django-sslserver==0.22
django-staticinline==1.3.1
django-storages==1.9.1
django-tastypie==0.14.3
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.15.2
future==0.18.2
hiredis==2.0.0
html5lib==1.1
hyperlink==20.0.1
idna==2.8
incremental==21.3.0
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mailchimp-marketing==3.0.40
mccabe==0.6.1
msgpack==1.0.0
numpy==1.20.2
oauthlib==3.1.0
ordered-set==4.0.2
#Pillow==9.0.0
Pillow==8.3.0
psycopg2-binary==2.8.6
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.9.9
pygram==0.0.13b0
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-http-client==3.2.7
python-memcached==1.59
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.5
reportlab==3.5.56
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sendgrid==6.6.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
stripe==2.54.0
toml==0.10.2
tornado==6.0.4
Twisted==21.7.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.24.3
validate-email==1.3
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
zope.interface==5.1.2
